margin to float32

+++ models/Mt4Account.go
@@ -7,6 +7,6 @@ type Mt4Account struct {
 	Server     string  `json:"server,omitempty"`
 	Balance    float32 `json:"balance"`
 	Equity     float32 `json:"equity,omitempty"`
-	Margin     int     `json:"margin,omitempty"`
+	Margin     float32 `json:"margin,omitempty"`
 	FreeMargin float32 `json:"freemargin,omitempty"`
 }
